使用了整合后的c库写了test（未测试）

+++ include/sys/types.h
@@ -22,8 +22,8 @@ typedef int mode_t;
 typedef int off_t;
 typedef int uid_t;
 typedef int nlink_t;
-typedef unsigned long long size_t,useconds_t;
-typedef long long ssize_t;
+typedef unsigned long size_t,useconds_t;
+typedef long ssize_t;
 
 
 #endif	/* _INC_TYPES */
